change order --> seqorder
order means many things to many people -- there is a possibility
that a later Pfastetter order or Strahler order may be added.
Also, the existing NHD/NWM datasets have and order field. So
this name change is to distinguish the order in this dataset as
one that is sequential. Maybe another, more precise name will
make sense after we have worked with it a bit more.